+++ collector/src/main/java/com/bitmovin/analytics/DefaultCollector.kt
@@ -40,7 +40,7 @@ abstract class DefaultCollector<TPlayer> protected constructor(
             // this is needed to mimic the existing behaviour and since we are
             // resetting the impressionId on attaching of the collector, which would mean that
             // the user gets a different impressionId before attaching
-            return if (analytics.playerAdapter != null) {
+            return if (analytics.isAttachedToPlayer()) {
                 analytics.impressionId
             } else {
                 null